Added utils folder with constants and common-functions file
