+++ packages/smart/hardhat.config.ts
@@ -6,7 +6,7 @@ import "hardhat-docgen";
 
 import { HardhatUserConfig } from "hardhat/config";
 import "./tasks";
-import { mapOverObject } from "./src";
+import { mapOverObject } from "./src/";
 import { NetworkUserConfig } from "hardhat/types";
 
 const ETHERSCAN_API_KEY = process.env["ETHERSCAN_API_KEY"] || "CH7M2ATCZABP2GIHEF3FREWWQPDFQBSH8G";
